ui: Allow changing c from the UI
Add a toggle button to the header bar displaying the current value of
c (for f(z) = z^2 + c) and show a popover when clicking on it, which
allows changing the values for the real and img part of c.

Note: When changing c all GdkPixbufs in the hashtable will be freed,
since they become useless and caching for all zoom level and c
combinations is not realistic.