animation: Use Clutter to animate zooming
The goals/constraints for this branch:
* always center at (0,0)
* use clutter to animate zoom level changes
* get this thread safe

The animation is done by scaling the actor with the current bitmap to
fit the new zoom level and then blending in another actor with the
newly fetched/created bitmap.

For big zoom jumps to look good, it would require multiple transitions
with bitmaps of intermediate zoom levels.