gmp: First attempt to use GMP
GMP is a free library for arbitrary precision arithmetic. See
https://gmplib.org/.

We can now achieve much larger zoom levels at the cost of dramatically
increased calculation times (compared to usual floating point
operations). This code is a first attempt and still has lots of
potential for optimizations. Also calculations currently block the
main thread UI.

This branch, as it is, performs too poorly to be considered
usable for now.